requirements.txt
flask
pytest
flask-sqlalchemy
flask-script
requests
pymysql
gunicorn
python-decouple
gevent
